Updated: rename tabular animation component to Loader
